requirements.txt
gensim
pyenchant
plotly
numpy
sklearn
nbformat